refactor: reorganize the reader dock into small, focused components
Splits the 731-line TheReaderDock.vue god component into a dock/
directory of single-purpose pieces (buttons, menus, pickers, a shared
Segmented control and MenuBox wrapper), dedupes repeated CSS/state,
fixes latent TS errors around the wake listener and idle timeout, and
moves icons/, App.vue, and index.scss under components/ and scss/ for
consistent organization.
