+++ src/readability/components/icons/IconAlignLeft.vue
