+++ src/readability/components/icons/IconClose.vue
