+++ src/readability/components/icons/IconCoffee.vue
