+++ src/readability/index.ts
@@ -1,6 +1,5 @@
-import './index.scss';
+import './scss/index.scss';
 
 export { apply, remove } from './apply';
 export { isReaderable } from './heuristics';
-export { readabilityFonts } from './fonts';
 export { SIZES, WIDTHS, nearestStepIndex } from './steps';
