+++ src/readability/scss/index.scss
@@ -7,7 +7,7 @@
 .stylebot-reader {
   --body-padding: 64px;
 
-  @import './scss/content.scss';
+  @import './content.scss';
 
   --main-background: #faf8f3;
   --main-foreground: #2b2926;
